Compact Script

+++ entity/file.go
@@ -1,7 +1,7 @@
 package entity
 
 type File struct {
-	Id   uint   `gorm:"primary_key" json:"id"`
+	Id   int64  `gorm:"primary_key" json:"id"`
 	Name string `gorm:"not null" json:"name" validate:"required"`
 	Path string `gorm:"not null" json:"path" validate:"required"`
 }
